Split implementation into multiple files.
Create modules for:
- iterator
- region
- span
