+++ src/libpmemstream.c
@@ -12,130 +12,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-static uint8_t *pmemstream_offset_to_ptr(struct pmemstream *stream, size_t offset)
-{
-	return (uint8_t *)stream->data->spans + offset;
-}
-
-static pmemstream_span_bytes *pmemstream_offset_to_span_ptr(struct pmemstream *stream, size_t offset)
-{
-	assert(offset % sizeof(pmemstream_span_bytes) == 0);
-
-	return (pmemstream_span_bytes *)pmemstream_offset_to_ptr(stream, offset);
-}
-
-static void pmemstream_span_create_empty(struct pmemstream *stream, uint64_t offset, size_t data_size)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	assert((data_size & PMEMSTREAM_SPAN_TYPE_MASK) == 0);
-	span[0] = data_size | PMEMSTREAM_SPAN_EMPTY;
-
-	void *dest = ((uint8_t *)span) + SPAN_EMPTY_METADATA_SIZE;
-	stream->memset(dest, 0, data_size, PMEM2_F_MEM_NONTEMPORAL | PMEM2_F_MEM_NODRAIN);
-	stream->persist(span, SPAN_EMPTY_METADATA_SIZE);
-}
-
-static void pmemstream_span_create_entry(struct pmemstream *stream, uint64_t offset, const void *data, size_t data_size,
-					 size_t popcount)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	assert((data_size & PMEMSTREAM_SPAN_TYPE_MASK) == 0);
-	span[0] = data_size | PMEMSTREAM_SPAN_ENTRY;
-	span[1] = popcount;
-
-	// XXX - use variadic mempcy to store data and metadata at once
-	void *dest = ((uint8_t *)span) + SPAN_ENTRY_METADATA_SIZE;
-	stream->memcpy(dest, data, data_size, PMEM2_F_MEM_NONTEMPORAL | PMEM2_F_MEM_NODRAIN);
-	stream->persist(span, SPAN_ENTRY_METADATA_SIZE);
-}
-
-static void pmemstream_span_create_region(struct pmemstream *stream, uint64_t offset, size_t size)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	assert((size & PMEMSTREAM_SPAN_TYPE_MASK) == 0);
-	span[0] = size | PMEMSTREAM_SPAN_REGION;
-
-	stream->persist(span, SPAN_REGION_METADATA_SIZE);
-}
-
-static uint64_t pmemstream_get_span_size(pmemstream_span_bytes *span)
-{
-	return span[0] & PMEMSTREAM_SPAN_EXTRA_MASK;
-}
-
-static enum pmemstream_span_type pmemstream_get_span_type(pmemstream_span_bytes *span)
-{
-	return span[0] & PMEMSTREAM_SPAN_TYPE_MASK;
-}
-
-static struct pmemstream_span_runtime pmemstream_span_get_empty_runtime(struct pmemstream *stream, uint64_t offset)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	struct pmemstream_span_runtime srt;
-
-	assert(pmemstream_get_span_type(span) == PMEMSTREAM_SPAN_EMPTY);
-
-	srt.type = PMEMSTREAM_SPAN_EMPTY;
-	srt.empty.size = pmemstream_get_span_size(span);
-	srt.data_offset = offset + SPAN_EMPTY_METADATA_SIZE;
-	srt.total_size = ALIGN_UP(srt.empty.size + SPAN_EMPTY_METADATA_SIZE, sizeof(pmemstream_span_bytes));
-
-	return srt;
-}
-
-static struct pmemstream_span_runtime pmemstream_span_get_entry_runtime(struct pmemstream *stream, uint64_t offset)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	struct pmemstream_span_runtime srt;
-
-	assert(pmemstream_get_span_type(span) == PMEMSTREAM_SPAN_ENTRY);
-
-	srt.type = PMEMSTREAM_SPAN_ENTRY;
-	srt.entry.size = pmemstream_get_span_size(span);
-	srt.entry.popcount = span[1];
-	srt.data_offset = offset + SPAN_ENTRY_METADATA_SIZE;
-	srt.total_size = ALIGN_UP(srt.entry.size + SPAN_ENTRY_METADATA_SIZE, sizeof(pmemstream_span_bytes));
-
-	return srt;
-}
-
-static struct pmemstream_span_runtime pmemstream_span_get_region_runtime(struct pmemstream *stream, uint64_t offset)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	struct pmemstream_span_runtime srt;
-
-	assert(pmemstream_get_span_type(span) == PMEMSTREAM_SPAN_REGION);
-
-	srt.type = PMEMSTREAM_SPAN_REGION;
-	srt.region.size = pmemstream_get_span_size(span);
-	srt.data_offset = offset + SPAN_REGION_METADATA_SIZE;
-	srt.total_size = ALIGN_UP(srt.region.size + SPAN_REGION_METADATA_SIZE, sizeof(pmemstream_span_bytes));
-
-	return srt;
-}
-
-static struct pmemstream_span_runtime pmemstream_span_get_runtime(struct pmemstream *stream, uint64_t offset)
-{
-	pmemstream_span_bytes *span = pmemstream_offset_to_span_ptr(stream, offset);
-	struct pmemstream_span_runtime srt;
-
-	switch (pmemstream_get_span_type(span)) {
-		case PMEMSTREAM_SPAN_EMPTY:
-			srt = pmemstream_span_get_empty_runtime(stream, offset);
-			break;
-		case PMEMSTREAM_SPAN_ENTRY:
-			srt = pmemstream_span_get_entry_runtime(stream, offset);
-			break;
-		case PMEMSTREAM_SPAN_REGION:
-			srt = pmemstream_span_get_region_runtime(stream, offset);
-			break;
-		default:
-			abort();
-	}
-
-	return srt;
-}
-
 static int pmemstream_is_initialized(struct pmemstream *stream)
 {
 	if (strcmp(stream->data->header.signature, PMEMSTREAM_SIGNATURE) != 0) {
@@ -164,17 +40,6 @@ static void pmemstream_init(struct pmemstream *stream)
 		       PMEM2_F_MEM_NONTEMPORAL);
 }
 
-static int validate_entry(struct pmemstream *stream, struct pmemstream_entry entry)
-{
-	struct pmemstream_span_runtime srt = pmemstream_span_get_runtime(stream, entry.offset);
-	void *entry_data = pmemstream_offset_to_ptr(stream, srt.data_offset);
-	if (srt.type == PMEMSTREAM_SPAN_ENTRY &&
-	    util_popcount_memory(entry_data, srt.entry.size) == srt.entry.popcount) {
-		return 0;
-	}
-	return -1;
-}
-
 int pmemstream_from_map(struct pmemstream **stream, size_t block_size, struct pmem2_map *map)
 {
 	struct pmemstream *s = malloc(sizeof(struct pmemstream));
@@ -193,14 +58,9 @@ int pmemstream_from_map(struct pmemstream **stream, size_t block_size, struct pm
 		pmemstream_init(s);
 	}
 
-	s->region_contexts_container = critnib_new();
-	if (!s->region_contexts_container) {
+	s->region_contexts_map = region_contexts_map_new();
+	if (!s->region_contexts_map) {
 		return -1;
-	}
-
-	int ret = pthread_mutex_init(&s->region_contexts_container_lock, NULL);
-	if (ret) {
-		return ret;
 	}
 
 	*stream = s;
@@ -208,22 +68,10 @@ int pmemstream_from_map(struct pmemstream **stream, size_t block_size, struct pm
 	return 0;
 }
 
-int critnib_free_context(uintptr_t key, void *value, void *privdata)
-{
-	free(value);
-	return 0;
-}
-
 void pmemstream_delete(struct pmemstream **stream)
 {
 	struct pmemstream *s = *stream;
-
-	critnib_iter(s->region_contexts_container, 0, (uint64_t)-1, critnib_free_context, NULL);
-	critnib_delete(s->region_contexts_container);
-
-	/* XXX: Handle error */
-	pthread_mutex_destroy(&s->region_contexts_container_lock);
-
+	region_contexts_map_destroy(s->region_contexts_map);
 	free(s);
 	*stream = NULL;
 }
@@ -262,8 +110,7 @@ int pmemstream_region_free(struct pmemstream *stream, struct pmemstream_region r
 
 	pmemstream_span_create_empty(stream, 0, stream->usable_size - SPAN_EMPTY_METADATA_SIZE);
 
-	struct pmemstream_region_context *ctx = critnib_remove(stream->region_contexts_container, region.offset);
-	free(ctx);
+	region_contexts_map_remove(stream->region_contexts_map, region);
 
 	return 0;
 }
@@ -284,245 +131,10 @@ size_t pmemstream_entry_length(struct pmemstream *stream, struct pmemstream_entr
 	return entry_srt.entry.size;
 }
 
-/* Gets (or creates if missing) pointer to region_context associated with specified region. */
-static int get_or_insert_region_context(struct pmemstream *stream, struct pmemstream_region region,
-					struct pmemstream_region_context **container_handle)
-{
-	struct pmemstream_region_context *ctx = critnib_get(stream->region_contexts_container, region.offset);
-	if (ctx) {
-		if (container_handle) {
-			*container_handle = ctx;
-		}
-		return 0;
-	}
-
-	int ret = -1;
-
-	pthread_mutex_lock(&stream->region_contexts_container_lock);
-	ctx = malloc(sizeof(*ctx));
-	if (ctx) {
-		*ctx = (struct pmemstream_region_context){0};
-		ret = critnib_insert(stream->region_contexts_container, region.offset, ctx, 0 /* no update */);
-	}
-	pthread_mutex_unlock(&stream->region_contexts_container_lock);
-
-	if (ret) {
-		/* Insert failed, free the context. */
-		free(ctx);
-
-		if (ret == EEXIST) {
-			/* Someone else inserted the region context - just get a pointer to it. */
-			ctx = critnib_get(stream->region_contexts_container, region.offset);
-			assert(ctx);
-		} else {
-			return ret;
-		}
-	}
-
-	if (container_handle) {
-		*container_handle = ctx;
-	}
-
-	return 0;
-}
-
-int pmemstream_region_iterator_new(struct pmemstream_region_iterator **iterator, struct pmemstream *stream)
-{
-	struct pmemstream_region_iterator *iter = malloc(sizeof(*iter));
-	iter->stream = stream;
-	iter->region.offset = 0;
-
-	*iterator = iter;
-
-	return 0;
-}
-
-int pmemstream_region_iterator_next(struct pmemstream_region_iterator *it, struct pmemstream_region *region)
-{
-	struct pmemstream_span_runtime srt;
-
-	while (it->region.offset < it->stream->usable_size) {
-		srt = pmemstream_span_get_runtime(it->stream, it->region.offset);
-
-		if (srt.type == PMEMSTREAM_SPAN_REGION) {
-			*region = it->region;
-			it->region.offset += srt.total_size;
-			return 0;
-		}
-
-		/* if there are no more regions we should expect an empty span */
-		assert(srt.type == PMEMSTREAM_SPAN_EMPTY);
-		it->region.offset += srt.total_size;
-	}
-
-	return -1;
-}
-
-void pmemstream_region_iterator_delete(struct pmemstream_region_iterator **iterator)
-{
-	struct pmemstream_region_iterator *iter = *iterator;
-
-	free(iter);
-	*iterator = NULL;
-}
-
-static int pmemstream_entry_iterator_initialize(struct pmemstream_entry_iterator *iterator, struct pmemstream *stream,
-						struct pmemstream_region region)
-{
-	struct pmemstream_span_runtime region_srt = pmemstream_span_get_region_runtime(stream, region.offset);
-	struct pmemstream_entry_iterator iter;
-	iter.offset = region_srt.data_offset;
-	iter.region = region;
-	iter.stream = stream;
-
-	int ret = get_or_insert_region_context(stream, region, &iter.region_context);
-	if (ret) {
-		return ret;
-	}
-
-	*iterator = iter;
-
-	return 0;
-}
-
-int pmemstream_entry_iterator_new(struct pmemstream_entry_iterator **iterator, struct pmemstream *stream,
-				  struct pmemstream_region region)
-{
-	struct pmemstream_entry_iterator *iter = malloc(sizeof(*iter));
-
-	int ret = pmemstream_entry_iterator_initialize(iter, stream, region);
-	if (ret) {
-		free(iter);
-		return ret;
-	}
-
-	*iterator = iter;
-
-	return 0;
-}
-
 int pmemstream_get_region_context(struct pmemstream *stream, struct pmemstream_region region,
 				  struct pmemstream_region_context **region_context)
 {
-	return get_or_insert_region_context(stream, region, region_context);
-}
-
-static int is_region_recovered(struct pmemstream_region_context *region_context)
-{
-	return __atomic_load_n(&region_context->append_offset, __ATOMIC_ACQUIRE) != PMEMSTREAM_OFFSET_UNINITIALIZED;
-}
-
-/* Iterates over entire region Might perform recovery. */
-static int region_iterate(struct pmemstream *stream, struct pmemstream_region region)
-{
-	struct pmemstream_entry_iterator iter;
-	int ret = pmemstream_entry_iterator_initialize(&iter, stream, region);
-	if (ret) {
-		return ret;
-	}
-
-	while (pmemstream_entry_iterator_next(&iter, NULL, NULL) == 0) {
-	}
-
-	return 0;
-}
-
-/* Recovers a region (under a global lock) if it is not yet recovered. */
-static int maybe_recover_region_locked(struct pmemstream *stream, struct pmemstream_region region,
-				       struct pmemstream_region_context *region_context)
-{
-	assert(region_context);
-	int ret = 0;
-
-	/* If region is not recovered, iterate over region and perform recovery.
-	 * Uses "double-checked locking". */
-	if (!is_region_recovered(region_context)) {
-		/* XXX: this can be per-region lock */
-		pthread_mutex_lock(&stream->region_contexts_container_lock);
-		if (!is_region_recovered(region_context)) {
-			ret = region_iterate(stream, region);
-		}
-		pthread_mutex_unlock(&stream->region_contexts_container_lock);
-	}
-
-	return ret;
-}
-
-/* Performs stream recovery - clears all the data in the region after `tail` entry. */
-static void recover_region(struct pmemstream *stream, struct pmemstream_region region,
-			   struct pmemstream_region_context *region_context, struct pmemstream_entry tail)
-{
-	assert(region_context);
-	assert(tail.offset != PMEMSTREAM_OFFSET_UNINITIALIZED);
-
-	struct pmemstream_span_runtime region_rt = pmemstream_span_get_region_runtime(stream, region.offset);
-	size_t region_end_offset = region.offset + region_rt.total_size;
-	size_t remaining_size = region_end_offset - tail.offset;
-
-	uint64_t expected_append_offset = PMEMSTREAM_OFFSET_UNINITIALIZED;
-	int weak = 0; /* Use compare_exchange int strong variation. */
-	if (__atomic_compare_exchange_n(&region_context->append_offset, &expected_append_offset, tail.offset, weak,
-					__ATOMIC_RELEASE, __ATOMIC_RELAXED)) {
-		pmemstream_span_create_empty(stream, tail.offset, remaining_size - SPAN_EMPTY_METADATA_SIZE);
-	} else {
-		/* Nothing to do, someone else already performed recovery. */
-	}
-}
-
-/* Advances entry iterator by one. Verifies entry integrity and recovers the region if necessary. */
-int pmemstream_entry_iterator_next(struct pmemstream_entry_iterator *iter, struct pmemstream_region *region,
-				   struct pmemstream_entry *user_entry)
-{
-	struct pmemstream_span_runtime srt = pmemstream_span_get_runtime(iter->stream, iter->offset);
-	struct pmemstream_span_runtime region_srt =
-		pmemstream_span_get_region_runtime(iter->stream, iter->region.offset);
-	struct pmemstream_entry entry;
-	entry.offset = iter->offset;
-
-	// XXX: add test for NULL entry
-	if (user_entry) {
-		*user_entry = entry;
-	}
-	if (region) {
-		*region = iter->region;
-	}
-
-	/* Make sure that we didn't go beyond region. */
-	if (iter->offset >= iter->region.offset + region_srt.total_size) {
-		return -1;
-	}
-
-	iter->offset += srt.total_size;
-
-	/*
-	 * Verify that all metadata and data fits inside the region - this should not fail unless stream was corrupted.
-	 */
-	assert(entry.offset + srt.total_size <= iter->region.offset + region_srt.total_size);
-
-	int region_recovered = is_region_recovered(iter->region_context);
-
-	if (region_recovered && srt.type == PMEMSTREAM_SPAN_EMPTY) {
-		/* If we found last entry and region is already recovered, just return -1. */
-		return -1;
-	} else if (!region_recovered && validate_entry(iter->stream, entry) < 0) {
-		/* If region was not yet recovered, validate that entry is correct. If there is any problem, recover the
-		 * region. */
-		recover_region(iter->stream, iter->region, iter->region_context, entry);
-		return -1;
-	}
-
-	/* Region is already recovered, and we did not enounter end of the data yet - span must be a valid entry */
-	assert(validate_entry(iter->stream, entry) == 0);
-
-	return 0;
-}
-
-void pmemstream_entry_iterator_delete(struct pmemstream_entry_iterator **iterator)
-{
-	struct pmemstream_entry_iterator *iter = *iterator;
-
-	free(iter);
-	*iterator = NULL;
+	return region_contexts_map_get_or_create(stream->region_contexts_map, region, region_context);
 }
 
 // synchronously appends data buffer to the end of the region
@@ -542,7 +154,7 @@ int pmemstream_append(struct pmemstream *stream, struct pmemstream_region region
 		}
 	}
 
-	ret = maybe_recover_region_locked(stream, region, region_context);
+	ret = region_try_recover_locked(stream, region, region_context);
 	if (ret) {
 		return ret;
 	}
